+++ lib/src/gallery/src/gallery_view.dart
@@ -78,9 +78,11 @@ class _GalleryViewWrapperState extends State<GalleryViewWrapper> {
     }
     final ps = _controller.panelSetting;
     final hs = _controller.headerSetting;
-    final _panelMaxHeight = ps.maxHeight ?? MediaQuery.of(context).size.height - hs.topMargin;
-    final _panelMinHeight = (ps.minHeight ?? kKeyboardHeight ?? _panelMaxHeight * 0.37) -
-        (widget.safeAreaBottom ? MediaQuery.of(context).padding.bottom : 0);
+    final _panelMaxHeight =
+        ps.maxHeight ?? MediaQuery.of(context).size.height - hs.topMargin;
+    final _panelMinHeight =
+        (ps.minHeight ?? kKeyboardHeight ?? _panelMaxHeight * 0.37) -
+            (widget.safeAreaBottom ? MediaQuery.of(context).padding.bottom : 0);
 
     final showKeyboard = MediaQuery.of(context).viewInsets.bottom != 0.0;
 
@@ -176,7 +178,8 @@ class GalleryView extends StatefulWidget {
   _GalleryViewState createState() => _GalleryViewState();
 }
 
-class _GalleryViewState extends State<GalleryView> with SingleTickerProviderStateMixin {
+class _GalleryViewState extends State<GalleryView>
+    with SingleTickerProviderStateMixin {
   late final GalleryController _controller;
   late final PanelController _panelController;
 
@@ -283,7 +286,8 @@ class _GalleryViewState extends State<GalleryView> with SingleTickerProviderStat
     }
 
     if (_controller.value.selectedEntities.isNotEmpty &&
-        (_controller.setting.onUnselectAll == null || _controller.setting.onUnselectAll!())) {
+        (_controller.setting.onUnselectAll == null ||
+            _controller.setting.onUnselectAll!())) {
       _showAlert(_controller.setting);
       return false;
     }
@@ -318,9 +322,12 @@ class _GalleryViewState extends State<GalleryView> with SingleTickerProviderStat
   Widget build(BuildContext context) {
     final ps = _controller.panelSetting;
     final hs = _controller.headerSetting;
-    final _panelMaxHeight = ps.maxHeight ?? MediaQuery.of(context).size.height - hs.topMargin;
-    final _panelMinHeight = ps.minHeight ?? kKeyboardHeight ?? _panelMaxHeight * 0.37;
-    final _setting = ps.copyWith(maxHeight: _panelMaxHeight, minHeight: _panelMinHeight);
+    final _panelMaxHeight =
+        ps.maxHeight ?? MediaQuery.of(context).size.height - hs.topMargin;
+    final _panelMinHeight =
+        ps.minHeight ?? kKeyboardHeight ?? _panelMaxHeight * 0.37;
+    final _setting =
+        ps.copyWith(maxHeight: _panelMaxHeight, minHeight: _panelMinHeight);
     final _headerSetting = hs;
 
     final albumListHeight = _panelMaxHeight - hs.headerMaxHeight;
@@ -349,7 +356,8 @@ class _GalleryViewState extends State<GalleryView> with SingleTickerProviderStat
               builder: (context) {
                 if (_controller.fullScreenMode) {
                   return SizedBox(
-                    height: _headerSetting.headerMaxHeight + MediaQuery.of(context).padding.top,
+                    height: _headerSetting.headerMaxHeight +
+                        MediaQuery.of(context).padding.top,
                   );
                 }
 
@@ -357,7 +365,10 @@ class _GalleryViewState extends State<GalleryView> with SingleTickerProviderStat
                   valueListenable: _panelController,
                   builder: (context, SliderValue value, child) {
                     final height = (_headerSetting.headerMinHeight +
-                            (_headerSetting.headerMaxHeight - _headerSetting.headerMinHeight) * value.factor * 1.2)
+                            (_headerSetting.headerMaxHeight -
+                                    _headerSetting.headerMinHeight) *
+                                value.factor *
+                                1.2)
                         .clamp(
                       _headerSetting.headerMinHeight,
                       _headerSetting.headerMaxHeight,
@@ -404,7 +415,9 @@ class _GalleryViewState extends State<GalleryView> with SingleTickerProviderStat
           animation: _animation,
           builder: (context, child) {
             final offsetY = _headerSetting.headerMaxHeight +
-                (_controller.fullScreenMode ? MediaQuery.of(context).padding.top : 0) +
+                (_controller.fullScreenMode
+                    ? MediaQuery.of(context).padding.top
+                    : 0) +
                 (_panelMaxHeight - hs.headerMaxHeight) * (1 - _animation.value);
             return Visibility(
               visible: _animation.value > 0.0,
@@ -559,7 +572,8 @@ class _GalleryViewFieldState extends State<GalleryViewField> {
           context,
         );
       },
-      child: widget.previewBuilder != null && (_controller.recentEntities?.isNotEmpty ?? false)
+      child: widget.previewBuilder != null &&
+              (_controller.recentEntities?.isNotEmpty ?? false)
           ? GalleryRecentPreview(
               entity: _controller.recentEntities!.first,
               builder: widget.previewBuilder,
@@ -673,7 +687,8 @@ class GalleryController extends ValueNotifier<GalleryValue> {
       value = const GalleryValue();
       return;
     }
-    final _afterRemove = value.selectedEntities..removeWhere((element) => list.contains(element));
+    final _afterRemove = value.selectedEntities
+      ..removeWhere((element) => list.contains(element));
     if (_afterRemove.isEmpty) {
       _onSubmitted?.call([]);
       _clearedSelection = true;
@@ -752,7 +767,9 @@ class GalleryController extends ValueNotifier<GalleryValue> {
   /// When panel closed without any selection
   void _closePanel() {
     _panelController.closePanel();
-    final entities = (_clearedSelection || value.selectedEntities.isEmpty) ? <LikkEntity>[] : value.selectedEntities;
+    final entities = (_clearedSelection || value.selectedEntities.isEmpty)
+        ? <LikkEntity>[]
+        : value.selectedEntities;
     _completer.complete(entities);
     // _onSubmitted?.call(entities);
     // _checkKeyboard.value = false;
@@ -854,7 +871,8 @@ class GalleryController extends ValueNotifier<GalleryValue> {
   }) async {
     // If dont have permission dont do anything
     final permission = await PhotoManager.requestPermissionExtend();
-    if (permission != PermissionState.authorized && permission != PermissionState.limited) {
+    if (permission != PermissionState.authorized &&
+        permission != PermissionState.limited) {
       PhotoManager.openSetting();
       return [];
     }
@@ -904,7 +922,8 @@ class GalleryController extends ValueNotifier<GalleryValue> {
   ///
   /// return true if selected media reached to maximum selection limit
   ///
-  bool get reachedMaximumLimit => value.selectedEntities.length == setting.maximum;
+  bool get reachedMaximumLimit =>
+      value.selectedEntities.length == setting.maximum;
 
   ///
   /// return true is gallery is in single selection mode